# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/auth/dorm-owner/login": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "login dorm owner",
                "parameters": [
                    {
                        "description": "login credentials",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LoginCredentialsDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.LoginDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/dorm-owner/register": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "register dorm owner account",
                "parameters": [
                    {
                        "description": "dorm owner registration",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RegisterDormOwnerDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.DormOwner"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/student/login": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "login student",
                "parameters": [
                    {
                        "description": "login credentials",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LoginCredentialsDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.LoginDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/student/register": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "register student account",
                "parameters": [
                    {
                        "description": "student registration",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RegisterStudentDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Student"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.LoginCredentialsDTO": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "dto.LoginDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
        "dto.RegisterDormOwnerDTO": {
            "type": "object",
            "required": [
                "email",
                "firstname",
                "lastname",
                "line_id",
                "password",
                "phone_number"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "line_id": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
...
        "dto.RegisterStudentDTO": {
            "type": "object",
            "required": [
                "email",
                "enrollment_year",
                "faculty",
                "firstname",
                "gender",
                "lastname",
                "password",
                "student_id"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "enrollment_year": {
                    "type": "integer"
                },
                "faculty": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "student_id": {
                    "type": "string"
                }
            }
        },
...
        "model.AllDormFacility": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "model.AllRoomFacility": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "model.BankAccount": {
            "type": "object",
            "properties": {
                "account_name": {
                    "type": "string"
                },
                "account_number": {
                    "type": "string"
                },
                "bank": {
                    "type": "string"
                },
                "bank_qr_url": {
                    "type": "string"
                }
            }
        },
        "model.Dorm": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "facilities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.AllDormFacility"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "lat": {
                    "type": "number"
                },
                "long": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "nearby_locations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.NearbyLocation"
                    }
                },
                "pictures": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.DormPicture"
                    }
                },
                "rooms": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Room"
                    }
                },
                "rules": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "zone": {
                    "type": "string"
                }
            }
        },
        "model.DormOwner": {
            "type": "object",
            "properties": {
                "bank_account": {
                    "$ref": "#/definitions/model.BankAccount"
                },
                "dorms": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Dorm"
                    }
                },
                "email": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastname": {
                    "type": "string"
                },
                "line_id": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "picture_url": {
                    "type": "string"
                }
            }
        },
        "model.DormPicture": {
            "type": "object",
            "properties": {
                "picture_url": {
                    "type": "string"
                }
            }
        },
...
        "model.Location": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "lat": {
                    "type": "number"
                },
                "long": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.NearbyLocation": {
            "type": "object",
            "properties": {
                "distance": {
                    "type": "number"
                },
                "location": {
                    "$ref": "#/definitions/model.Location"
                }
            }
        },
        "model.PersonalHabit": {
            "type": "object",
            "properties": {
                "pet_habit": {
                    "$ref": "#/definitions/model.PetHabit"
                },
                "room_care_habit": {
                    "$ref": "#/definitions/model.RoomCareHabit"
                },
                "sleep_habit": {
                    "$ref": "#/definitions/model.SleepHabit"
                },
                "smoke_habit": {
                    "$ref": "#/definitions/model.SmokeHabit"
                },
                "study_habit": {
                    "$ref": "#/definitions/model.StudyHabit"
                }
            }
        },
        "model.PetHabit": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.PetPicture": {
            "type": "object",
            "properties": {
                "picture_url": {
                    "type": "string"
                }
            }
        },
...
        "model.Room": {
            "type": "object",
            "properties": {
                "available_from": {
                    "type": "string"
                },
                "capacity": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "dorm_id": {
                    "type": "integer"
                },
                "facilities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.AllRoomFacility"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "pictures": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.RoomPicture"
                    }
                },
                "price": {
                    "type": "integer"
                },
                "size": {
                    "type": "number"
                }
            }
        },
        "model.RoomCareHabit": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.RoomPicture": {
            "type": "object",
            "properties": {
                "picture_url": {
                    "type": "string"
                }
            }
        },
        "model.RoommatePreference": {
            "type": "object",
            "properties": {
                "other_preference": {
                    "type": "string"
                },
                "preferred_gender": {
                    "type": "string"
                },
                "preferred_pet_habit": {
                    "$ref": "#/definitions/model.PetHabit"
                },
                "preferred_room_care_habit": {
                    "$ref": "#/definitions/model.RoomCareHabit"
                },
                "preferred_sleep_habit": {
                    "$ref": "#/definitions/model.SleepHabit"
                },
                "preferred_smoke_habit": {
                    "$ref": "#/definitions/model.SmokeHabit"
                },
                "preferred_study_habit": {
                    "$ref": "#/definitions/model.StudyHabit"
                }
            }
        },
...
        "model.SleepHabit": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.SmokeHabit": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.SocialMedia": {
            "type": "object",
            "properties": {
                "facebook_url": {
                    "type": "string"
                },
                "instagram_url": {
                    "type": "string"
                },
                "linkedin_url": {
                    "type": "string"
                },
                "twitter_url": {
                    "type": "string"
                }
            }
        },
        "model.Student": {
            "type": "object",
            "properties": {
                "biography": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "enrollment_year": {
                    "type": "integer"
                },
                "faculty": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "personal_habit": {
                    "$ref": "#/definitions/model.PersonalHabit"
                },
                "pet_pictures": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.PetPicture"
                    }
                },
                "picture_url": {
                    "type": "string"
                },
                "roommate_preference": {
                    "$ref": "#/definitions/model.RoommatePreference"
                },
                "social_media": {
                    "$ref": "#/definitions/model.SocialMedia"
                }
            }
        },
        "model.StudyHabit": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}